fix(fe): refactor save
